requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
click==7.0                # via flask
flask==1.0.2
itsdangerous==0.24        # via flask
jinja2==2.10              # via flask
markupsafe==1.0           # via jinja2
rpi.bme280==0.2.2
smbus2==0.2.1
werkzeug==0.14.1          # via flask
